refactor!: split Employee.FederalTaxes into onboarding + management variants
Replaces the `isOnboarding` prop with two journey-scoped components that share
the same form rendering but differ in CTAs and emitted events:

- `onboarding/FederalTaxes.tsx` — Continue button, emits
  `EMPLOYEE_FEDERAL_TAXES_UPDATED` + `EMPLOYEE_FEDERAL_TAXES_DONE`.
- `management/FederalTaxes.tsx` — Cancel + Save, dismissible success alert,
  emits `CANCEL` or `EMPLOYEE_FEDERAL_TAXES_UPDATED` (no `_DONE`).
- `shared/FederalTaxesView.tsx` — heading, intro, fields, and slot props
  (`actions`, `alert`, `children` override) consumed by both variants.

`Employee.FederalTaxes` now resolves to the management variant (matches the
WorkAddress/HomeAddress convention and the prior `isOnboarding=false`
default). `EmployeeOnboarding.FederalTaxes` and `EmployeeManagement.FederalTaxes`
each point at their dedicated variant.

Consumers rewired:
- OnboardingFlow / SelfOnboardingFlow → onboarding variant (no more
  `isOnboarding` prop).
- Dashboard → management variant.

Tests split into `onboarding/FederalTaxes.test.tsx` (4 tests) and
`management/FederalTaxes.test.tsx` (6 tests). Workflow docs updated to
describe the two variants instead of the prop toggle.
