+++ src/components/Employee/FederalTaxes/management/index.ts
@@ -1,0 +1,1 @@
+export { FederalTaxes, type FederalTaxesProps } from './FederalTaxes'
