+++ src/components/Employee/FederalTaxes/shared/index.ts
@@ -1,0 +1,2 @@
+export { FederalTaxesView, type FederalTaxesViewProps } from './FederalTaxesView'
+export * from './useFederalTaxesForm'
