+++ src/components/Employee/exports/employeeManagement.ts
@@ -3,7 +3,7 @@ export { EmployeeDocuments } from '../EmployeeDocuments'
 export { DashboardFlow } from '../Dashboard'
 export { WorkAddress } from '../WorkAddress/management/WorkAddress'
 export type { WorkAddressProps } from '../WorkAddress/management/WorkAddress'
-export { FederalTaxes, type FederalTaxesProps } from '../FederalTaxes'
+export { FederalTaxes, type FederalTaxesProps } from '../FederalTaxes/management/FederalTaxes'
 export { TerminateEmployee } from '../Terminations/TerminateEmployee/TerminateEmployee'
 export { TerminationSummary } from '../Terminations/TerminationSummary/TerminationSummary'
 export { TerminationFlow } from '../Terminations/TerminationFlow/TerminationFlow'
