+++ src/components/Employee/index.ts
@@ -4,6 +4,7 @@ export { OnboardingSummary } from './OnboardingSummary'
 export { Profile } from './Profile'
 export { Compensation } from './Compensation'
 export { FederalTaxes } from './FederalTaxes'
+export type { FederalTaxesProps } from './FederalTaxes'
 export { StateTaxes } from './StateTaxes'
 export { PaymentMethod } from './PaymentMethod'
 export { Landing } from './Landing'
